Exported EA function stack
